Renamed Logging.hpp to LoggingFramework.hpp.
